feat(homepage): ✨ create homepage with blocks
