+++ apps/registry/app/docs/layout.tsx
@@ -1,4 +1,5 @@
 import { DocsSidebarNav } from "@/components/docs-sidebar";
+import { Navbar } from "@/components/navbar";
 import registry from "@/registry.json";
 
 export default function DocsLayout({
@@ -44,13 +45,16 @@ export default function DocsLayout({
 	];
 
 	return (
-		<div className="relative mx-auto w-full max-w-[540px] px-4 md:max-w-[720px] lg:max-w-[960px] xl:max-w-[1140px]">
-			<div className="grid grid-cols-1 gap-8 py-8 md:grid-cols-[250px_1fr]">
-				<aside className="border-r pr-4">
-					<DocsSidebarNav items={items} />
-				</aside>
-				<main className="prose max-w-none dark:prose-invert">{children}</main>
+		<>
+			<Navbar />
+			<div className="relative mx-auto w-full max-w-[540px] px-4 md:max-w-[720px] lg:max-w-[960px] xl:max-w-[1140px]">
+				<div className="grid grid-cols-1 gap-8 py-8 md:grid-cols-[250px_1fr]">
+					<aside className="border-r pr-4">
+						<DocsSidebarNav items={items} />
+					</aside>
+					<main className="prose max-w-none dark:prose-invert">{children}</main>
+				</div>
 			</div>
-		</div>
+		</>
 	);
 }
