+++ apps/registry/components/dropdown-menu-gameplan.tsx
@@ -1,4 +1,3 @@
-import { Button } from "@/components/ui/button";
 import {
 	DropdownMenu,
 	DropdownMenuContent,
